Update 2018-01-03-season-1-episode-1.markdown
Modification of the link to the MP3 file on archive.org to track it with the Blubrry stat counter and reverting the link for the rss feed to its original location. Blubrry tracks each files through a specially crafted link instead of an RSS redirect. We can thus keep our actual RSS feed generated by Jekyll.